split tracing docs into rails and other frameworks
