+++ docs-content/getting-started/6_backend-tracing/6_ruby/1_overview.md
@@ -1,0 +1,20 @@
+---
+toc: Overview
+title: Tracing in Ruby
+slug: overview
+---
+
+Highlight.io supports tracing in Ruby.
+
+```hint
+If you don't see one of your languages / frameworks below, reach out to us in our [community](https://highlight.io/community) or create an issue on [github](https://github.com/highlight/highlight/issues/new?assignees=&labels=external+bug+%2F+request&template=feature_request.md&title=).
+```
+
+<DocsCardGroup>
+    <DocsCard title="Rails" href="./rails.md">
+        {"Integrate tracing in Rails."}
+    </DocsCard>
+    <DocsCard title="Other Ruby Frameworks" href="./other.md">
+        {"Integrate tracing in other Ruby Frameworks."}
+    </DocsCard>
+</DocsCardGroup>
