+++ docs-content/getting-started/6_backend-tracing/6_ruby/2_rails.md
@@ -1,7 +1,7 @@
 ---
-toc: Ruby
-title: Backend Tracing in Ruby
-slug: ruby
+toc: Rails
+title: Rails
+slug: rails
 quickstart: true
 ---
