+++ docs-content/getting-started/6_backend-tracing/6_ruby/3_other.md
@@ -1,0 +1,8 @@
+---
+toc: Ruby App
+title: Tracing with Other Ruby Frameworks
+slug: other
+quickstart: true
+---
+
+<QuickStart content={quickStartContent["traces"]["ruby"]["other"]}/>
