+++ docs-content/getting-started/6_backend-tracing/6_ruby/index.md
@@ -1,0 +1,5 @@
+---
+toc: Ruby
+title: Backend Tracing in Ruby
+slug: ruby
+---
