+++ highlight.io/components/QuickstartContent/traces/ruby/other.tsx
@@ -1,12 +1,12 @@
-import { siteUrl } from '../../../utils/urls'
-import { initializeSdk, installSdk } from '../backend/ruby/shared-snippets'
-import { frontendInstallSnippet } from '../backend/shared-snippets'
-import { QuickStartContent } from '../QuickstartContent'
+import { siteUrl } from '../../../../utils/urls'
+import { initializeSdk, installSdk } from '../../backend/ruby/shared-snippets'
+import { frontendInstallSnippet } from '../../backend/shared-snippets'
+import { QuickStartContent } from '../../QuickstartContent'
+import { verifyTraces } from '../shared-snippets'
 
-export const RubyTracesContent: QuickStartContent = {
+export const RubyOtherTracesContent: QuickStartContent = {
 	title: 'Ruby',
-	subtitle:
-		'Learn how to set up highlight.io on your non-Rails Ruby backend.',
+	subtitle: 'Learn how to set up highlight.io tracing for Ruby applications.',
 	logoUrl: siteUrl('/images/quickstart/ruby.svg'),
 	entries: [
 		frontendInstallSnippet,
@@ -28,5 +28,6 @@ end`,
 				},
 			],
 		},
+		verifyTraces,
 	],
 }
